chore: abstract some reusable pieces
